r flytter 12 uker egenvurdering inn i 14a onboaridngen
