--- src/komponenter/12uker-egenvurdering/12uker-egenvurdering.less
@@ -1,61 +1,0 @@
-@import '../../../node_modules/nav-frontend-core/less/_variabler';
-
-.egenvurdering-wrapper {
-  background: #0067C5;
-  color: #ffffff;
-  margin: 0.5rem -1rem;
-}
-
-.egenvurdering-innhold {
-  display: flex;
-}
-
-.egenvurdering-avslaa-knapp {
-  background: transparent;
-  color: black;
-  cursor: pointer;
-  border: 0;
-  height: 32px;
-  width: 32px;
-}
-
-.egenvurdering-avslaa-knapp:hover {
-  background: #0067C5;
-  color: white;
-}
-
-.p-1 {
-  padding: 1rem;
-}
-
-.flex {
-  display: flex;
-}
-
-.space-between {
-  justify-content: space-between;
-}
-
-.flex-column {
-  flex-direction: column;
-}
-
-@media (min-width: @screen-md-min) {
-  .tolv-uker-intro-omslutning {
-    .panel {
-      border-bottom: 0;
-      height: 100%;
-    }
-  }
-  .tolv-uker-intro-omslutning-modal {
-    margin: 0;
-    flex: 1;
-    .panel {
-      background: none;
-      border-radius: 0;
-      border: 0 solid @navGra20;
-      height: 48ch;
-      width: 50ch;
-    }
-  }
-}
