+++ src/komponenter/onboarding14a/14a-intro.less
@@ -1,27 +1,21 @@
 @import '../../../node_modules/nav-frontend-core/less/_variabler';
 
-.fjorten-A-intro-omslutning-modal {
-    margin: 0;
-    flex: 1;
-    .panel {
-        background: none;
-        border-radius: 0;
-        border: 0 solid @navGra20;
-        height: 60ch;
-        width: 80vw;
+.egenvurdering-wrapper {
+    background: #0067c5;
+    color: #ffffff;
+    margin: 0.5rem -1rem;
+    padding: 1rem;
+    .egenvurdering-avslaa-knapp {
+        background: transparent;
+        color: black;
+        cursor: pointer;
+        border: 0;
+        height: 32px;
+        width: 32px;
+    }
+    .egenvurdering-avslaa-knapp:hover {
+        background: #0067c5;
+        color: white;
     }
 }
-
-@media (min-width: @screen-md-min) {
-    .fjorten-A-intro-omslutning-modal {
-        margin: 0;
-        flex: 1;
-        .panel {
-            background: none;
-            border-radius: 0;
-            border: 0 solid @navGra20;
-            height: 48ch;
-            width: 50ch;
-        }
-    }
-}
+
